src/test/javascript/spec/app/entities/result-additional-info/result-additional-info.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { ResultAdditionalInfoService } from 'app/entities/result-additional-info/result-additional-info.service';
import { IResultAdditionalInfo, ResultAdditionalInfo } from 'app/shared/model/result-additional-info.model';

describe('Service Tests', () => {
  describe('ResultAdditionalInfo Service', () => {
    let injector: TestBed;
    let service: ResultAdditionalInfoService;
    let httpMock: HttpTestingController;
    let elemDefault: IResultAdditionalInfo;
    let expectedResult: IResultAdditionalInfo | IResultAdditionalInfo[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(ResultAdditionalInfoService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new ResultAdditionalInfo(0, 'image/png', 'AAAAAAA', 'image/png', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a ResultAdditionalInfo', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new ResultAdditionalInfo()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a ResultAdditionalInfo', () => {
        const returnedFromService = Object.assign(
          {
            gpxTrack: 'BBBBBB',
            heartRate: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of ResultAdditionalInfo', () => {
        const returnedFromService = Object.assign(
          {
            gpxTrack: 'BBBBBB',
            heartRate: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a ResultAdditionalInfo', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
